Finish and polish up thread and cpp/thread API

+++ include/os/mqx/etcpal/os_thread.h
@@ -50,7 +50,7 @@ typedef struct
   _task_id tid;
 } etcpal_thread_t;
 
-#define etcpal_thread_sleep(sleep_ms) _time_delay(sleep_ms)
+#define etcpal_thread_sleep(sleep_ms) _time_delay((uint32_t)sleep_ms)
 
 #ifdef __cplusplus
 }
